refactor(#467): extract usePhaseGameCompletion hook, apply to 4 game hooks
