+++ gamesformykids/hooks/shared/progress/index.ts
@@ -3,3 +3,4 @@ export { useSessionStats } from './useSessionStats';
 export { useAchievements } from './useAchievements';
 export { useGameCompletion } from './useGameCompletion';
 export type { GameResult } from './useGameCompletion';
+export { usePhaseGameCompletion } from './usePhaseGameCompletion';
